ps2 keyboard translating scancodes to machine-universal keycodes

+++ kernel/text_mode.c
@@ -68,7 +68,7 @@ void print_string(char* str, uint8_t col) {
 }
 
 void print_hex32(uint32_t x, uint8_t col) {
-    int32_t a = x;
+    uint32_t a = x;
     for (int32_t i = 0; i < 8; i++) {
         int8_t half_byte_val = ((a >> 28) & 0x0f);
         char c;
@@ -102,7 +102,7 @@ void print_memory_block(uint32_t from, uint32_t to) {
 
     while ((uint32_t)memptr < to) {
         print_char('\n', 0);
-        print_hex32(memptr, TM_FORE_COL_BLUE);
+        print_hex32((uint32_t)memptr, TM_FORE_COL_BLUE);
         print_string("    ", 0);
         
         for (uint8_t j = 0; j < 16; j++) {
